WIP Share code between app and server to create the notification content
