--- firebase-debug.log
@@ -1,22 +1,0 @@
-[debug] [2025-04-20T22:22:28.349Z] ----------------------------------------------------------------------
-[debug] [2025-04-20T22:22:28.352Z] Command:       C:\Program Files\nodejs\node.exe C:\Users\Martin\AppData\Roaming\npm\node_modules\firebase-tools\lib\bin\firebase.js deploy --only functions
-[debug] [2025-04-20T22:22:28.352Z] CLI Version:   14.2.0
-[debug] [2025-04-20T22:22:28.353Z] Platform:      win32
-[debug] [2025-04-20T22:22:28.353Z] Node Version:  v20.19.0
-[debug] [2025-04-20T22:22:28.353Z] Time:          Sun Apr 20 2025 23:22:28 GMT+0100 (British Summer Time)
-[debug] [2025-04-20T22:22:28.353Z] ----------------------------------------------------------------------
-[debug]
-[debug] [2025-04-20T22:22:28.614Z] > command requires scopes: ["email","openid","https://www.googleapis.com/auth/cloudplatformprojects.readonly","https://www.googleapis.com/auth/firebase","https://www.googleapis.com/auth/cloud-platform"]
-[debug] [2025-04-20T22:22:28.615Z] > authorizing via signed-in user (martin.ter.haak@gmail.com)
-[debug] [2025-04-20T22:22:28.615Z] [iam] checking project whatshouldiweartoday-e1a4b for permissions ["cloudfunctions.functions.create","cloudfunctions.functions.delete","cloudfunctions.functions.get","cloudfunctions.functions.list","cloudfunctions.functions.update","cloudfunctions.operations.get","firebase.projects.get"]
-[debug] [2025-04-20T22:22:28.616Z] Checked if tokens are valid: true, expires at: 1745189695545
-[debug] [2025-04-20T22:22:28.617Z] Checked if tokens are valid: true, expires at: 1745189695545
-[debug] [2025-04-20T22:22:28.618Z] >>> [apiv2][query] POST https://cloudresourcemanager.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b:testIamPermissions [none]
-[debug] [2025-04-20T22:22:28.618Z] >>> [apiv2][(partial)header] POST https://cloudresourcemanager.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b:testIamPermissions x-goog-quota-user=projects/whatshouldiweartoday-e1a4b
-[debug] [2025-04-20T22:22:28.618Z] >>> [apiv2][body] POST https://cloudresourcemanager.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b:testIamPermissions {"permissions":["cloudfunctions.functions.create","cloudfunctions.functions.delete","cloudfunctions.functions.get","cloudfunctions.functions.list","cloudfunctions.functions.update","cloudfunctions.operations.get","firebase.projects.get"]}
-[debug] [2025-04-20T22:22:29.337Z] <<< [apiv2][status] POST https://cloudresourcemanager.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b:testIamPermissions 200
-[debug] [2025-04-20T22:22:29.338Z] <<< [apiv2][body] POST https://cloudresourcemanager.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b:testIamPermissions {"permissions":["cloudfunctions.functions.create","cloudfunctions.functions.delete","cloudfunctions.functions.get","cloudfunctions.functions.list","cloudfunctions.functions.update","cloudfunctions.operations.get","firebase.projects.get"]}
-[debug] [2025-04-20T22:22:29.338Z] Checked if tokens are valid: true, expires at: 1745189695545
-[debug] [2025-04-20T22:22:29.338Z] Checked if tokens are valid: true, expires at: 1745189695545
-[debug] [2025-04-20T22:22:29.338Z] >>> [apiv2][query] POST https://iam.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b/serviceAccounts/whatshouldiweartoday-e1a4b@appspot.gserviceaccount.com:testIamPermissions [none]
-[debug] [2025-04-20T22:22:29.338Z] >>> [apiv2][body] POST https://iam.googleapis.com/v1/projects/whatshouldiweartoday-e1a4b/serviceAccounts/whatshouldiweartoday-e1a4b@appspot.gserviceaccount.com:testIamPermissions {"permissions":["iam.serviceAccounts.actAs"]}
