+++ functions/.gitignore
@@ -1,3 +1,4 @@
 node_modules/
 *.local
 build/
+.env.*
