+++ functions/package.json
@@ -3,7 +3,8 @@
   "description": "Cloud Functions for Firebase",
   "scripts": {
     "lint": "eslint .",
-    "build": "tsc",
+    "build": "tsc --build",
+    "build:watch": "tsc --build --watch",
     "serve": "npm run build && firebase emulators:start --only functions",
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
@@ -13,7 +14,7 @@
   "engines": {
     "node": "18"
   },
-  "main": "build/src/index.js",
+  "main": "build/functions/src/index.js",
   "dependencies": {
     "date-fns-tz": "^3.2.0",
     "expo-server-sdk": "^3.7.0",
