+++ functions/tsconfig.json
@@ -4,13 +4,23 @@
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "outDir": "build",
-    "rootDir": ".",
     "sourceMap": true,
     "strict": true,
     "target": "es2017",
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "composite": true,
+    "baseUrl": ".",
+    "rootDir": ".."
   },
   "compileOnSave": true,
-  "include": ["src"],
-  "exclude": ["node_modules", "build"]
+  "include": [
+    "src/**/*",
+    "../shared/src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "build"
+  ]
 }
