+++ package.json
@@ -1,5 +1,10 @@
 {
   "name": "whatshouldiweartoday",
+  "private": true,
+  "workspaces": [
+    "functions",
+    "shared"
+  ],
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
