+++ shared/src/types/Time.ts
@@ -1,0 +1,4 @@
+export interface Time {
+    hours: number;
+    minutes: number;
+}
