+++ shared/src/types/WeatherForecast.ts
