+++ src/__tests__/mockData.ts
@@ -1,8 +1,8 @@
 import Location from "../models/Location";
 import Time from "../models/Time";
-import UserProfile from "../models/UserProfile";
-import WeatherForecast from "../models/WeatherForecast";
-import { Gender, TemperatureUnit } from "../models/enums";
+import UserProfile from "../../shared/types/UserProfile";
+import WeatherForecast from "../../shared/types/WeatherForecast";
+import { Gender, TemperatureUnit } from "../../shared/types/enums";
 
 export const mockNavigation = {
   navigate: jest.fn(),
