+++ src/services/Store.ts
@@ -6,7 +6,7 @@ import { httpsCallable } from "firebase/functions";
 import { auth, functions } from "../config/firebase";
 import Time from "../models/Time";
 import UserProfile from "../models/UserProfile";
-import { Gender, TemperatureUnit } from "../models/enums";
+import { Gender, TemperatureUnit } from "../../shared/src/types/enums";
 
 const INITIAL_PROFILE = new UserProfile(
   null,
