+++ tsconfig.json
@@ -10,5 +10,9 @@
     "strict": true
   },
   "extends": "expo/tsconfig.base",
-  "include": ["**/*.ts", "**/*.tsx"]
+  "include": ["**/*.ts", "**/*.tsx"],
+  "files": [],
+  "references": [
+    { "path": "./shared" }
+  ]
 }
